@@ -15,7 +15,7 @@ type StorageConfig =
 
 module MemoryStore =
     type [<NoEquality; NoComparison>] Arguments =
-        | [<AltCommandLine("-vs")>] VerboseStore
+        | [<AltCommandLine "-V">]        VerboseStore
         interface IArgParserTemplate with
             member a.Usage = a |> function
                 | VerboseStore ->       "Include low level Store logging."
@@ -34,8 +34,8 @@ module Cosmos =
         | Some x -> x
 
     type [<NoEquality; NoComparison>] Arguments =
-        | [<AltCommandLine "-vs">]      VerboseStore
-        | [<AltCommandLine "-m">]       ConnectionMode of Equinox.Cosmos.ConnectionMode
+        | [<AltCommandLine "-V">]       VerboseStore
+        | [<AltCommandLine "-m">]       ConnectionMode of Microsoft.Azure.Cosmos.ConnectionMode
         | [<AltCommandLine "-o">]       Timeout of float
         | [<AltCommandLine "-r">]       Retries of int
         | [<AltCommandLine "-rt">]      RetriesWaitTimeS of float
@@ -54,7 +54,7 @@ module Cosmos =
                 | Database _ ->         "specify a database name for store. (optional if environment variable EQUINOX_COSMOS_DATABASE specified)"
                 | Container _ ->        "specify a container name for store. (optional if environment variable EQUINOX_COSMOS_CONTAINER specified)"
     type Info(args : ParseResults<Arguments>) =
-        member __.Mode =                args.GetResult(ConnectionMode,Equinox.Cosmos.ConnectionMode.Direct)
+        member __.Mode =                args.GetResult(ConnectionMode,Microsoft.Azure.Cosmos.ConnectionMode.Direct)
         member __.Connection =          args.TryGetResult Connection |> defaultWithEnvVar "EQUINOX_COSMOS_CONNECTION" "Connection"
         member __.Database =            args.TryGetResult Database   |> defaultWithEnvVar "EQUINOX_COSMOS_DATABASE"   "Database"
         member __.Container =           args.TryGetResult Container  |> defaultWithEnvVar "EQUINOX_COSMOS_CONTAINER"  "Container"
@@ -89,7 +89,7 @@ module Cosmos =
 ///   2. & $env:ProgramData\chocolatey\bin\EventStore.ClusterNode.exe --gossip-on-single-node --discover-via-dns 0 --ext-http-port=30778
 module EventStore =
     type [<NoEquality; NoComparison>] Arguments =
-        | [<AltCommandLine("-vs")>]     VerboseStore
+        | [<AltCommandLine("-V")>]      VerboseStore
         | [<AltCommandLine("-o")>]      Timeout of float
         | [<AltCommandLine("-r")>]      Retries of int
         | [<AltCommandLine("-g")>]      Host of string
