@@ -1,15 +1,14 @@
 ﻿module Backend.Favorites
 
+open Domain
 open Domain.Favorites
 open System
 
-type Service(log, resolve, ?maxAttempts) =
-
-    let resolve (Events.ForClientId streamId) = Equinox.Stream(log, resolve streamId, defaultArg maxAttempts 2)
+type Service internal (resolve : ClientId -> Equinox.Stream<Events.Event, Fold.State>) =
 
     let execute clientId command : Async<unit> =
         let stream = resolve clientId
-        stream.Transact(Commands.interpret command)
+        stream.Transact(interpret command)
     let read clientId : Async<Events.Favorited []> =
         let stream = resolve clientId
         stream.Query id
@@ -20,8 +19,12 @@ type Service(log, resolve, ?maxAttempts) =
     member __.Favorite(clientId, skus) =
         execute clientId (Command.Favorite(DateTimeOffset.Now, skus))
 
-    member __.Unfavorite(clientId, skus) =
-        execute clientId (Command.Unfavorite skus)
+    member __.Unfavorite(clientId, sku) =
+        execute clientId (Command.Unfavorite sku)
 
     member __.List clientId : Async<Events.Favorited []> =
-        read clientId 
+        read clientId
+
+let create log resolve =
+    let resolve id = Equinox.Stream(log, resolve (streamName id), maxAttempts  = 3)
+    Service(resolve)
