@@ -4,15 +4,13 @@ open Domain
 open Domain.SavedForLater
 open System
 
-type Service(handlerLog, resolve, maxSavedItems : int, ?maxAttempts) =
+type Service internal (resolve : ClientId -> Equinox.Stream<Events.Event, Fold.State>, maxSavedItems) =
 
     do if maxSavedItems < 0 then invalidArg "maxSavedItems" "must be non-negative value."
 
-    let resolve (Events.ForClientId streamId) = Equinox.Stream(handlerLog, resolve streamId, defaultArg maxAttempts 3)
-
     let execute clientId command : Async<bool> =
         let stream = resolve clientId
-        stream.Transact(Commands.decide maxSavedItems command)
+        stream.Transact(decide maxSavedItems command)
     let read clientId : Async<Events.Item[]> =
         let stream = resolve clientId
         stream.Query id
@@ -21,7 +19,7 @@ type Service(handlerLog, resolve, maxSavedItems : int, ?maxAttempts) =
         stream.TransactAsync(fun (state : Fold.State) -> async {
             let contents = seq { for item in state -> item.skuId } |> set
             let! cmd = resolveCommand contents.Contains
-            let _, events = Commands.decide maxSavedItems cmd state
+            let _, events = decide maxSavedItems cmd state
             return (),events } )
 
     member __.MaxSavedItems = maxSavedItems
@@ -40,3 +38,7 @@ type Service(handlerLog, resolve, maxSavedItems : int, ?maxAttempts) =
     member __.Merge(clientId, targetId) : Async<bool> = async {
         let! state = read clientId
         return! execute targetId (Merge state) }
+
+let create maxSavedItems log resolve =
+    let resolve id = Equinox.Stream(log, resolve (streamName id), maxAttempts = 3)
+    Service(resolve, maxSavedItems)
