@@ -1,9 +1,11 @@
 ﻿module Domain.ContactPreferences
 
 type Id = Id of email: string
+let streamName (Id email) = FsCodec.StreamName.create "ContactPreferences" email // TODO hash >> base64
 
 // NOTE - these types and the union case names reflect the actual storage formats and hence need to be versioned with care
 module Events =
+
     type Preferences = { manyPromotions : bool; littlePromotions : bool; productReview : bool; quickSurveys : bool }
     type Value = { email : string; preferences : Preferences }
 
@@ -11,9 +13,9 @@ module Events =
         | [<System.Runtime.Serialization.DataMember(Name = "contactPreferencesChanged")>]Updated of Value
         interface TypeShape.UnionContract.IUnionContract
     let codec = FsCodec.NewtonsoftJson.Codec.Create<Event>()
-    let (|ForClientId|) (email: string) = Equinox.AggregateId ("ContactPreferences", email) // TODO hash >> base64
 
 module Fold =
+
     type State = Events.Preferences
 
     let initial : State = { manyPromotions = false; littlePromotions = false; productReview = false; quickSurveys = false }
@@ -29,9 +31,8 @@ module Fold =
 type Command =
     | Update of Events.Value
 
-module Commands =
-    let interpret command (state : Fold.State) =
-        match command with
-        | Update ({ preferences = preferences } as value) ->
-            if state = preferences then [] else
-            [ Events.Updated value ] 
+let interpret command (state : Fold.State) =
+    match command with
+    | Update ({ preferences = preferences } as value) ->
+        if state = preferences then [] else
+        [ Events.Updated value ]
