@@ -72,14 +72,14 @@ module RequestId =
 /// CartId strongly typed id; represented internally as a Guid; not used for storage so rendering is not significant
 type CartId = Guid<cartId>
 and [<Measure>] cartId
-module CartId = let toStringN (value : CartId) : string = Guid.toStringN %value
+module CartId = let toString (value : CartId) : string = Guid.toStringN %value
 
 /// ClientId strongly typed id; represented internally as a Guid; not used for storage so rendering is not significant
 type ClientId = Guid<clientId>
 and [<Measure>] clientId
-module ClientId = let toStringN (value : ClientId) : string = Guid.toStringN %value
+module ClientId = let toString (value : ClientId) : string = Guid.toStringN %value
 
 /// InventoryItemId strongly typed id
 type InventoryItemId = Guid<inventoryItemId>
 and [<Measure>] inventoryItemId
-module InventoryItemId = let toStringN (value : InventoryItemId) : string = Guid.toStringN %value
+module InventoryItemId = let toString (value : InventoryItemId) : string = Guid.toStringN %value
