@@ -3,8 +3,11 @@ module Domain.InventoryItem
 
 open System
 
+let streamName (id : InventoryItemId) = FsCodec.StreamName.create "InventoryItem" (InventoryItemId.toString id)
+
 // NOTE - these types and the union case names reflect the actual storage formats and hence need to be versioned with care
 module Events =
+
     type Event =
         | Created of name: string
         | Deactivated
@@ -12,7 +15,6 @@ module Events =
         | Removed of count: int
         | CheckedIn of count: int
         interface TypeShape.UnionContract.IUnionContract
-    let (|ForInventoryItemId|) (id : InventoryItemId) = Equinox.AggregateId ("InventoryItem", InventoryItemId.toStringN id)
 
 module Fold =
     type State = { active : bool; name: string; quantity: int }
@@ -33,24 +35,23 @@ type Command =
     | CheckIn of count: int
     | Deactivate
 
-module Commands =
-    // TODO make commands/event representations idempotent
-    let interpret command (state : Fold.State) =
-        match command with
-        | Create name ->
-            if String.IsNullOrEmpty name then invalidArg "name" ""
-            if state.name = name then [] else
-            [ Events.Created name ]
-        | ChangeName newName ->
-            if String.IsNullOrEmpty newName then invalidArg "newName" ""
-            if state.name = newName then [] else
-            [ Events.Renamed newName ]
-        | Remove count ->
-            if count <= 0 then invalidOp "cant remove negative count from inventory"
-            [ Events.Removed count]
-        | CheckIn count ->
-            if count <= 0 then invalidOp "must have a count greater than 0 to add to inventory"
-            [ Events.CheckedIn count ]
-        | Deactivate ->
-            if not state.active then invalidOp "Already deactivated"
-            [ Events.Deactivated ] 
+// TODO make commands/event representations idempotent
+let interpret command (state : Fold.State) =
+    match command with
+    | Create name ->
+        if String.IsNullOrEmpty name then invalidArg "name" ""
+        if state.name = name then [] else
+        [ Events.Created name ]
+    | ChangeName newName ->
+        if String.IsNullOrEmpty newName then invalidArg "newName" ""
+        if state.name = newName then [] else
+        [ Events.Renamed newName ]
+    | Remove count ->
+        if count <= 0 then invalidOp "cant remove negative count from inventory"
+        [ Events.Removed count]
+    | CheckIn count ->
+        if count <= 0 then invalidOp "must have a count greater than 0 to add to inventory"
+        [ Events.CheckedIn count ]
+    | Deactivate ->
+        if not state.active then invalidOp "Already deactivated"
+        [ Events.Deactivated ]
