@@ -4,6 +4,20 @@ module Sequence
 
 open System
 
+let [<Literal>] Category = "Sequence"
+let streamName id = FsCodec.StreamName.create Category (SequenceId.toString id)
+
+// shim for net461
+module Seq =
+    let tryLast (source : seq<_>) =
+        use e = source.GetEnumerator()
+        if e.MoveNext() then
+            let mutable res = e.Current
+            while (e.MoveNext()) do res <- e.Current
+            Some res
+        else
+            None
+
 // NOTE - these types and the union case names reflect the actual storage formats and hence need to be versioned with care
 module Events =
 
@@ -12,8 +26,6 @@ module Events =
         | Reserved of Reserved
         interface TypeShape.UnionContract.IUnionContract
     let codec = FsCodec.NewtonsoftJson.Codec.Create<Event>()
-    let [<Literal>] categoryId = "Sequence"
-    let (|ForSequenceId|) id = Equinox.AggregateId(categoryId, SequenceId.toString id)
 
 module Fold =
 
@@ -28,32 +40,34 @@ module Fold =
 let decideReserve (count : int) (state : Fold.State) : int64 * Events.Event list =
     state.next,[Events.Reserved { next = state.next + int64 count }]
 
-type Service internal (log, resolve, maxAttempts) =
-
-    let resolve (Events.ForSequenceId streamId) = Equinox.Stream(log, resolve streamId, maxAttempts)
+type Service internal (resolve : SequenceId -> Equinox.Stream<Events.Event, Fold.State>) =
 
     /// Reserves an id, yielding the reserved value. Optional <c>count</c> enables reserving more than the default count of <c>1</c> in a single transaction
     member __.Reserve(series,?count) : Async<int64> =
         let stream = resolve series
         stream.Transact(decideReserve (defaultArg count 1))
 
-let create resolve = Service(Serilog.Log.ForContext<Service>(), resolve, maxAttempts = 3)
+let create resolve =
+    let resolve sequenceId =
+        let streamName = streamName sequenceId
+        Equinox.Stream(Serilog.Log.ForContext<Service>(), resolve streamName, maxAttempts = 3)
+    Service(resolve)
 
 module Cosmos =
 
     open Equinox.Cosmos
-    let private createService (context,cache,accessStrategy) =
+    let private create (context,cache,accessStrategy) =
         let cacheStrategy = CachingStrategy.SlidingWindow (cache, TimeSpan.FromMinutes 20.) // OR CachingStrategy.NoCaching
-        let resolve = Resolver(context, Events.codec, Fold.fold, Fold.initial, cacheStrategy, accessStrategy).Resolve
-        create resolve
+        let resolver = Resolver(context, Events.codec, Fold.fold, Fold.initial, cacheStrategy, accessStrategy)
+        create resolver.Resolve
 
     module LatestKnownEvent =
 
-        let createService (context,cache) =
+        let create (context,cache) =
             let accessStrategy = AccessStrategy.LatestKnownEvent
-            createService (context,cache,accessStrategy)
+            create (context,cache,accessStrategy)
 
     module RollingUnfolds =
 
-        let createService (context,cache) =
-            createService (context,cache,AccessStrategy.RollingState Fold.snapshot)
+        let create (context,cache) =
+            create (context,cache,AccessStrategy.RollingState Fold.snapshot)
