@@ -1,8 +1,8 @@
 /// Simple example of how one might have multiple uploaders agree/share a common UploadId for a given OrderId
 module Upload
 
-open System
 open FSharp.UMX
+open System
 
 type PurchaseOrderId = int<purchaseOrderId>
 and [<Measure>] purchaseOrderId
@@ -13,6 +13,20 @@ type CompanyId = string<companyId>
 and [<Measure>] companyId
 module CompanyId =
     let toString (value : CompanyId) : string = %value
+
+let [<Literal>] Category = "Upload"
+let streamName (companyId, purchaseOrderId) = FsCodec.StreamName.compose Category [PurchaseOrderId.toString purchaseOrderId; CompanyId.toString companyId]
+
+// shim for net461
+module Seq =
+    let tryLast (source : seq<_>) =
+        use e = source.GetEnumerator()
+        if e.MoveNext() then
+            let mutable res = e.Current
+            while (e.MoveNext()) do res <- e.Current
+            Some res
+        else
+            None
 
 type UploadId = string<uploadId>
 and [<Measure>] uploadId
@@ -27,10 +41,6 @@ module Events =
         | IdAssigned of IdAssigned
         interface TypeShape.UnionContract.IUnionContract
     let codec = FsCodec.NewtonsoftJson.Codec.Create<Event>()
-    let [<Literal>] categoryId = "Upload"
-    let (|ForCompanyAndPurchaseOrder|) (companyId, purchaseOrderId) =
-        let id = sprintf "%s_%s" (PurchaseOrderId.toString purchaseOrderId) (CompanyId.toString companyId)
-        Equinox.AggregateId(categoryId, id)
 
 module Fold =
 
@@ -39,33 +49,35 @@ module Fold =
     let private evolve _ignoreState = function
         | Events.IdAssigned e -> Some e.value
     let fold (state: State) (events: seq<Events.Event>) : State =
-        Seq.tryLast events |> Option.fold evolve state
+        events |> Seq.tryLast |> Option.fold evolve state
 
 let decide (value : UploadId) (state : Fold.State) : Choice<UploadId,UploadId> * Events.Event list =
     match state with
     | None -> Choice1Of2 value, [Events.IdAssigned { value = value}]
     | Some value -> Choice2Of2 value, []
 
-type Service internal (log, resolve, maxAttempts) =
-
-    let resolve (Events.ForCompanyAndPurchaseOrder streamId) = Equinox.Stream(log, resolve streamId, maxAttempts)
+type Service internal (resolve : CompanyId * PurchaseOrderId -> Equinox.Stream<Events.Event, Fold.State>) =
 
     member __.Sync(companyId, purchaseOrderId, value) : Async<Choice<UploadId,UploadId>> =
         let stream = resolve (companyId, purchaseOrderId)
         stream.Transact(decide value)
 
-let create resolve = Service(Serilog.Log.ForContext<Service>(), resolve, 3)
+let create resolve =
+    let resolve ids =
+        let streamName = streamName ids
+        Equinox.Stream(Serilog.Log.ForContext<Service>(), resolve streamName, maxAttempts = 3)
+    Service(resolve)
 
 module Cosmos =
 
     open Equinox.Cosmos
-    let createService (context,cache) =
+    let create (context,cache) =
         let cacheStrategy = CachingStrategy.SlidingWindow (cache, TimeSpan.FromMinutes 20.) // OR CachingStrategy.NoCaching
-        let resolve = Resolver(context, Events.codec, Fold.fold, Fold.initial, cacheStrategy, AccessStrategy.LatestKnownEvent).Resolve
-        create resolve
+        let resolver = Resolver(context, Events.codec, Fold.fold, Fold.initial, cacheStrategy, AccessStrategy.LatestKnownEvent)
+        create resolver.Resolve
 
 module EventStore =
     open Equinox.EventStore
-    let createService context =
-        let resolve = Resolver(context, Events.codec, Fold.fold, Fold.initial, access=AccessStrategy.LatestKnownEvent).Resolve
-        create resolve
+    let create context =
+        let resolver = Resolver(context, Events.codec, Fold.fold, Fold.initial, access=AccessStrategy.LatestKnownEvent)
+        create resolver.Resolve
