@@ -15,7 +15,7 @@ type GetByIdArgsTemplate = { id: int }
 
 // Fulfills contract dictated by https://www.todobackend.com
 // To run:
-//     & dotnet run -f netcoreapp2.1 -p samples/TodoBackend
+//     & dotnet run -f netcoreapp3.1 -p samples/TodoBackend
 //     https://www.todobackend.com/client/index.html?https://localhost:5001/todos
 // Stolen from https://github.com/ChristianAlexander/dotnetcore-todo-webapi/blob/master/src/TodoWebApi/Controllers/TodosController.cs
 // even more stolen from https://github.com/joeaudette/playground/blob/master/spa-stack/src/FSharp.WebLib/Controllers.fs
@@ -23,9 +23,9 @@ type GetByIdArgsTemplate = { id: int }
 type TodosController(service: Service) =
     inherit ControllerBase()
 
-    let toModel (value : TodoView) : Todo = { id = value.id; order = value.order; title = value.title; completed = value.completed }
+    let toModel (value : TodoView) : Events.Todo = { id = value.id; order = value.order; title = value.title; completed = value.completed }
 
-    member private __.WithUri(x : Todo) : TodoView =
+    member private __.WithUri(x : Events.Todo) : TodoView =
         let url = __.Url.RouteUrl("GetTodo", { id=x.id }, __.Request.Scheme) // Supplying scheme is secret sauce for making it absolute as required by client
         { id = x.id; url = url; order = x.order; title = x.title; completed = x.completed }
 
